fixed #316 ,code optimisation
