+++ tabs/grepResult.py
@@ -8,99 +8,22 @@
 grep検索の結果が格納されます。同じファイルで複数のヒットがあった場合、エントリの数が増えていきます。
 """
 
-import os
-import gettext
-import logging
-import wx
-import errorCodes
-import lists
-import browsableObjects
 import globalVars
-import fileOperator
-import misc
-import workerThreads
-import workerThreadTasks
-import StringUtil
+import lists
 
-from win32com.shell import shell, shellcon
-from . import fileList
+from . import searchResultTabBase
 
-class GrepResultTab(fileList.FileListTab):
+
+class GrepResultTab(searchResultTabBase.SearchResultTabBase):
 	"""grepの検索結果が表示されているタブ。"""
 
-	blockMenuList=[
-		"FILE_MKDIR",
-		"EDIT_PAST",
-		"EDIT_SEARCH",
+	blockMenuList=searchResultTabBase.SearchResultTabBase.blockMenuList+[
 		"MOVE_FORWARD_TAB",
 		"MOVE_TOPFILE",
-		"MOVE_BACKWARD",
-		"MOVE_MARKSET",
-		"MOVE_MARK",
-		"VIEW_DRIVE_INFO",
-		"TOOL_ADDPATH",
-		"TOOL_EJECT_DRIVE",
-		"TOOL_EJECT_DEVICE",
-		"TOOL_EXEC_PROGRAM"
 	]
 
-	def StartSearch(self,rootPath,searches,keyword, isRegularExpression):
-		self.listObject=lists.GrepResultList()
-		self.listObject.Initialize(rootPath,searches,keyword, isRegularExpression)
-		self.SetListColumns(self.listObject)
-		workerThreads.RegisterTask(workerThreadTasks.PerformSearch,{'listObject': self.listObject, 'tabObject': self})
-
-		#タブの名前変更を通知
-		globalVars.app.hMainView.UpdateTabName()
-
-	def _onSearchHitCallback(self,hits):
-		"""コールバックで、ヒットしたオブジェクトのリストが降ってくるので、それをリストビューに追加していく。"""
-		globalVars.app.PlaySound("click.ogg")
-		for elem in hits:
-			self.hListCtrl.Append(elem.GetListTuple())
-
-	def GoBackward(self):
-		return errorCodes.BOUNDARY
-
-	def GoForward(self,stream,admin=False):
-		"""検索結果表示では、フォルダを開くときに別タブを生成する。"""
-		index=self.GetFocusedItem()
-		elem=self.listObject.GetElement(index)
-		if (not stream) and (type(elem)==browsableObjects.File or type(elem)==browsableObjects.GrepItem) :#このファイルを開く
-			misc.RunFile(elem.fullpath,admin)
-			return
-		else:
-			#新しいタブで開く
-			globalVars.app.hMainView.Navigate(elem.fullpath,as_new_tab=True)
-		#end ファイルを開くか移動するか
-	#end GoForward
-
-	def ReadCurrentFolder(self):
-		state=_("grep検索完了") if self.listObject.GetFinishedStatus() is True else _("grep検索中")
-		globalVars.app.say(_("キーワード %(keyword)s で ディレクトリ %(dir)s を%(state)s") % {'keyword': self.listObject.GetKeywordString(), 'dir': self.listObject.rootDirectory, 'state': state}, interrupt=True)
-
-	def ReadListItemNumber(self,short=False):
-		globalVars.app.say(_("検索結果 %(results)d件") % {'results': len(self.listObject)}, interrupt=True)
+	#内部で利用するリストの種類を定義
+	listType=lists.GrepResultList
 
 	def ReadListInfo(self):
-		globalVars.app.say(_("%(keyword)sのgrep検索結果を %(sortkind)sの%(sortad)sで一覧中、 %(max)d個中 %(current)d個目") %{'keyword': self.listObject.GetKeywordString(), 'sortkind': self.listObject.GetSortKindString(), 'sortad': self.listObject.GetSortAdString(), 'max': len(self.listObject), 'current': self.GetFocusedItem()+1}, interrupt=True)
-
-	def UpdateFilelist(self,silence=False,cursorTargetName=""):
-		"""検索をやり直す。ファイルは非同期処理で増えるので、cursorTargetNameは使用されない。"""
-		if not self.listObject.GetFinishedStatus():
-			globalVars.app.say(_("現在検索中です。再建策はできません。"), interrupt=True)
-			return
-		#end まだ検索終わってない
-		if silence==False:
-			globalVars.app.say(_("再建策"), interrupt=True)
-		#end not silence
-		self.hListCtrl.DeleteAllItems()
-		self.listObject.RedoSearch()
-		workerThreads.RegisterTask(workerThreadTasks.PerformSearch,{'listObject': self.listObject, 'tabObject': self})
-
-	def GetTabName(self):
-		"""タブコントロールに表示する名前"""
-		word=self.listObject.GetKeywordString()
-		word=StringUtil.GetLimitedString(word,globalVars.app.config["view"]["header_title_length"])
-		return _("%(word)sの検索") % {"word":word}
-
+		globalVars.app.say(_("%(keyword)sのファイル内容検索結果を %(sortkind)sの%(sortad)sで一覧中、 %(max)d個中 %(current)d個目") %{'keyword': self.listObject.GetKeywordString(), 'sortkind': self.listObject.GetSortKindString(), 'sortad': self.listObject.GetSortAdString(), 'max': len(self.listObject), 'current': self.GetFocusedItem()+1}, interrupt=True)
